Move git_ui closer to editor
