+++ crates/panel/Cargo.toml
@@ -12,7 +12,6 @@ workspace = true
 path = "src/panel.rs"
 
 [dependencies]
-editor.workspace = true
 gpui.workspace = true
 settings.workspace = true
 theme.workspace = true
